Refactoring
Refactoring

+++ Business/Handlers/Groups/Commands/DeleteGroupCommand.cs
@@ -29,7 +29,13 @@ namespace Business.Handlers.Groups.Commands
             [LogAspect(typeof(FileLogger))]
             public async Task<IResult> Handle(DeleteGroupCommand request, CancellationToken cancellationToken)
             {
-                var groupToDelete = await _groupRepository.GetAsync(x => x.Id == request.Id);
+                var groupToDelete = await _groupRepository
+                    .GetAsync(x => x.Id == request.Id);
+
+                if (groupToDelete == null)
+                {
+                    return new ErrorResult(Messages.GroupNotFound);
+                }
 
                 _groupRepository.Delete(groupToDelete);
                 await _groupRepository.SaveChangesAsync();
