+++ Business/Handlers/OperationClaims/Commands/DeleteOperationClaimCommand.cs
@@ -30,6 +30,12 @@ namespace Business.Handlers.OperationClaims.Commands
             public async Task<IResult> Handle(DeleteOperationClaimCommand request, CancellationToken cancellationToken)
             {
                 var claimToDelete = await _operationClaimRepository.GetAsync(x => x.Id == request.Id);
+
+                if (claimToDelete == null)
+                {
+                    return new ErrorResult(Messages.OperationClaimNotFound);
+                }
+
                 _operationClaimRepository.Delete(claimToDelete);
                 await _operationClaimRepository.SaveChangesAsync();
